+++ scraper/db_settings.py.carp
@@ -6,5 +6,3 @@ mysql_port = 3306
 mysql_pw = 'changeme'
 
 site_url = 'https://www---smt-import-20170710-vt6syoq-3faqimkzadf6s.us.platform.sh/'
-
-execution_path = '/home/colin/smt-scrapy'
