+++ scraper/db_settings.py.colin
@@ -6,5 +6,3 @@ mysql_port = 30002
 mysql_pw = None
 
 site_url = 'http://www.socialmediatoday.com/'
-
-execution_path = '/Users/colin/Repositories/smt-scrapy'
