+++ scraper/utils.py
@@ -1,121 +1,4 @@
-import db_settings
-import MySQLdb
-import logging
-import json
 import string
-import codecs
-
-
-class AbstractNodeData(object):
-
-    def __init__(self, nodes_filename='nodes.json', urls_filename='urls.txt', encoding='latin-1'):
-        self._data = None
-        self._urls = None
-        self.nodes_filename = nodes_filename
-        self.urls_filename = urls_filename
-        self.encoding = encoding
-
-    def get_data(self):
-        if self._data is None:
-            self._data = self._query_data()
-
-        return self._data
-
-    def get_urls(self):
-        if self._urls is None:
-            self._urls = self._query_urls()
-
-        return self._urls
-
-    def _query_data(self):
-        return {}
-
-    def _query_urls(self):
-        return []
-
-
-class NodeDataReader(AbstractNodeData):
-
-    def _query_data(self):
-        with codecs.open(self.nodes_filename, 'rb') as fp:
-            all_node_data = json.load(fp, encoding=self.encoding)
-
-        return all_node_data
-
-    def _query_urls(self):
-        with codecs.open(self.urls_filename, 'rb', encoding=self.encoding) as fp:
-            urls = fp.readlines()
-
-        return [x.strip() for x in urls]
-
-
-class NodeDataWriter(AbstractNodeData):
-
-    def write_nodes_json(self):
-        with codecs.open(self.nodes_filename, 'wb') as fp:
-            json.dump(self.get_data(), fp, encoding=self.encoding)
-
-    def write_urls_txt_file(self):
-        with codecs.open(self.urls_filename, 'wb', encoding=self.encoding) as fp:
-            fp.write('\n'.join(self.get_urls()))
-
-    def _get_connection(self):
-        """ Establish a mysql database connection from given db_settings file values
-            Return the database connection object
-        """
-        kwargs = {}
-        if db_settings.mysql_pw:
-            kwargs['passwd'] = db_settings.mysql_pw
-        if db_settings.mysql_user:
-            kwargs['user'] = db_settings.mysql_user
-        if db_settings.mysql_host:
-            kwargs['host'] = db_settings.mysql_host
-        if db_settings.mysql_db:
-            kwargs['db'] = db_settings.mysql_db
-        if db_settings.mysql_port:
-            kwargs['port'] = db_settings.mysql_port
-
-        logging.debug("Connecting to mysql: %s" % kwargs)
-        db = MySQLdb.connect(**kwargs)
-        return db
-
-    def _query_data(self):
-        # this query gets back the identifying information for the posts we want
-        # as well as the node and full path URLs that we can use to get the current
-        # actual HTML page
-        db = self._get_connection()
-        node_query = (
-            "select n.nid, user.uid, user.mail, n.changed, "
-            "n.type as content_type, u.alias, u.source "
-            "from node n "
-            "join users user on n.uid = user.uid "
-            "join url_alias u on u.source = CONCAT('node/', n.nid) "
-            "where n.type ='post' and n.status = 1 "
-            "and u.pid=(select pid from url_alias where source = u.source "
-            "order by pid desc limit 1) "
-            "order by n.changed DESC"
-        )
-
-        node_cursor = db.cursor()
-        node_cursor.execute(node_query)
-        nodes = {}
-        for nid, uid, mail, changed, content_type, alias, source in node_cursor:
-            node_data = {
-                'nid': nid,
-                'uid': uid,
-                'user_email': mail,
-                'changed': changed,
-                'content_type': content_type,
-                'node_url_path': source,
-                'url_path': alias
-            }
-            nodes['%s%s' % (db_settings.site_url, alias)] = node_data
-        db.close()
-
-        return nodes
-
-    def _query_urls(self):
-        return self.get_data().keys()
 
 
 # page parsing utils
